Fixed minor warnings

+++ Platformer/Src/GUI/UI.cpp
@@ -9,7 +9,7 @@ void UI::Update(const sf::RenderWindow & main_window)
 {
 	for (Heart& obj : m_heartsArray)
 	{
-		obj.sprite.setPosition(main_window.mapPixelToCoords(sf::Vector2i(WindowWidth - (35 * (3 - obj.id)), 10.f)));
+		obj.sprite.setPosition(main_window.mapPixelToCoords(sf::Vector2i(WindowWidth - (35 * (3 - obj.id)), 10)));
 	}
 }
 
@@ -29,7 +29,7 @@ void UI::InitHealth()
 	{
 		obj.id = heart_id_count;
 		obj.sprite.setTexture(m_heartTexture);
-		obj.sprite.setPosition(sf::Vector2f(WindowWidth - (16 * (3 - heart_id_count)), 10.f));
+		obj.sprite.setPosition(sf::Vector2f(float(WindowWidth - (16 * (3 - heart_id_count))), 10.f));
 		obj.isActive = true;
 
 		++heart_id_count;
